split up filestream
